requirements.txt
fastapi~=0.115.12
uvicorn

pydantic~=2.11.5
PyYAML~=6.0.2
SQLAlchemy~=2.0.41